backend/requirements.txt
langchain-mcp-adapters
langchain-google-genai
langgraph
langchain
langchain-community
google-search-results
googlemaps
streamlit
requests
python-dotenv
mcp
fastapi[standard]
astrapy
langchain_astradb